major refactoring of artemis
